First pass at getting the credential manager to handle oauth refresh. Auto-test work. Doing a manual check now
