+++ src/container.ts
@@ -91,7 +91,7 @@ export class Container {
         this._context = context;
         this._version = version;
 
-        context.subscriptions.push((this._credentialManager = new CredentialManager(this._analyticsClient)));
+        context.subscriptions.push((this._credentialManager = new CredentialManager(context, this._analyticsClient)));
         context.subscriptions.push((this._siteManager = new SiteManager(context.globalState)));
         context.subscriptions.push((this._clientManager = new ClientManager(context)));
         context.subscriptions.push((this._jiraProjectManager = new JiraProjectManager()));
